formalised and partitioned code
