+++ startcp/codechef.py
@@ -1,13 +1,13 @@
 import os
-from pathlib import Path
-import shutil
 import re
 import requests
 
 try:
-    import printer, constants
+    import printer
+    import constants
+    import utilities
 except Exception:
-    from startcp import printer, constants
+    from startcp import printer, constants, utilities
 
 
 rangebi = printer.Rangebi()
@@ -62,63 +62,18 @@ def prepare_for_codechef_battle(problem_urls, comp_url):
 
             response = response.json()
 
-            if not os.path.isfile(problem_folder_name + "/" + "problem.html"):
-                html_str = get_java_script_code_for_problem(comp_url + "/problems/" + problem_url.split("/")[-1])
+            utilities.create_problem_html_file(
+                problem_folder_name + "/" + "problem.html",
+                comp_url + "/problems/" + problem_url.split("/")[-1]
+            )
 
-                problem_html_file = problem_folder_name + "/" + "problem.html"
-
-                with open(problem_html_file, "w+") as outfile:
-                    outfile.write(html_str)
-
-            tmplt_file_created = True
-            if (not (os.getenv(constants.use_template) is None)) and (int(os.getenv(constants.use_template)) == 1):
-                try:
-                    if not (os.getenv(constants.main_lang_template_path) is None):
-                        if Path(os.getenv(constants.main_lang_template_path)).is_file():
-                            shutil.copy(os.getenv(constants.main_lang_template_path), problem_folder_name + "/")
-                            if not (os.getenv(constants.backup_lang_template_path) is None):
-                                if Path(os.getenv(constants.backup_lang_template_path)).is_file():
-                                    shutil.copy(os.getenv(constants.backup_lang_template_path), problem_folder_name + "/")
-                        else:
-                            tmplt_file_created = False
-                    else:
-                        tmplt_file_created = False
-                except Exception:
-                    tmplt_file_created = False
-            else:
-                tmplt_file_created = False
-
-            if not tmplt_file_created:
-                if not os.path.isfile(problem_folder_name + "/" + "sol.py"):
-                    Path(problem_folder_name + "/" + "sol.py").touch()
-
-                if not os.path.isfile(problem_folder_name + "/" + "sol.cpp"):
-                    Path(problem_folder_name + "/" + "sol.cpp").touch()
+            utilities.create_solution_prog_files(problem_folder_name)
 
             for sample_test_case in response["problemComponents"]["sampleTestCases"]:
-
                 id = sample_test_case["id"]
                 input_str = sample_test_case["input"]
                 output_str = sample_test_case["output"]
 
-                input_filename = problem_folder_name + "/" + "in" + str(id) + ".txt"
-                output_filename = problem_folder_name + "/" + "out" + str(id) + ".txt"
-
-                with open(input_filename, "w+") as outfile:
-                    outfile.write(input_str)
-                with open(output_filename, "w+") as outfile:
-                    outfile.write(output_str)
+                utilities.create_input_output_files(problem_folder_name, input_str, output_str, id)
 
         problem_counter += 1
-
-
-def get_java_script_code_for_problem(problem_url):
-    return """
-    <html>
-        <body>
-            <script>
-                window.location.replace('{problem_url}');
-            </script>
-        </body>
-    </html>
-    """.format(problem_url=problem_url)
